refactor(shared): reorganize package with barrel exports
- Create schemas/index.ts barrel export for all Zod schemas
- Create types/index.ts barrel export for pure TypeScript types
- Reduce main index.ts from 340 lines to 70 lines
- Add ARCHITECTURE.md with module structure documentation
- Add missing exports.ts schema re-exports

Resolves #4
