+++ shared/src/types/index.ts
@@ -1,0 +1,19 @@
+/**
+ * @autoart/shared - Types Module
+ *
+ * Pure TypeScript type definitions (no Zod schemas).
+ * These are interface/type definitions only.
+ */
+
+// ==================== PROJECTIONS ====================
+export type {
+    ProjectionPreset,
+    ActionProjectionInput,
+    ContainerInput,
+    StageProjectionOutput,
+    HierarchyProjectionOutput,
+    ProcessProjectionOutput,
+    TimelineProjectionOutput,
+    ProjectionRef,
+    ProjectionSelection,
+} from '../projections.js';
